+++ DB/DBZ/standard/service/src/main/java/jp/co/ndensan/reams/db/dbz/service/core/basic/ImageManager.java
@@ -14,7 +14,6 @@ import jp.co.ndensan.reams.db.dbz.entity.basic.DbT5115ImageEntity;
 import jp.co.ndensan.reams.db.dbz.persistence.basic.DbT5115ImageDac;
 import jp.co.ndensan.reams.ur.urz.definition.enumeratedtype.message.UrSystemErrorMessages;
 import jp.co.ndensan.reams.uz.uza.biz.Code;
-import jp.co.ndensan.reams.uz.uza.math.Decimal;
 import jp.co.ndensan.reams.uz.uza.util.di.InstanceProvider;
 import jp.co.ndensan.reams.uz.uza.util.di.Transaction;
 
@@ -52,7 +51,7 @@ public class ImageManager {
     @Transaction
     public Image getイメージ情報(
             ShinseishoKanriNo 申請書管理番号,
-            Decimal 取込ページ番号,
+            int 取込ページ番号,
             Code 原本マスク分) {
         requireNonNull(申請書管理番号, UrSystemErrorMessages.値がnull.getReplacedMessage("申請書管理番号"));
         requireNonNull(取込ページ番号, UrSystemErrorMessages.値がnull.getReplacedMessage("取込ページ番号"));
